requirements.txt
# Core Dependencies
databricks-sql-connector>=3.0.0
databricks-connect>=13.0.0
databricks-cli>=0.18.0

# ML and AI
databricks-vectorsearch>=0.1.0
mlflow>=2.8.0
langchain>=0.1.0
langchain-community>=0.0.20
langchain-databricks>=0.1.0

# LLM and Embeddings
openai>=1.10.0
tiktoken>=0.5.0

# Data Processing
pyspark>=3.5.0
pandas>=2.1.0
numpy>=1.24.0
delta-spark>=3.0.0

# Configuration and Utilities
pyyaml>=6.0.1
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Azure SDK
azure-identity>=1.15.0
azure-keyvault-secrets>=4.7.0
azure-storage-blob>=12.19.0
azure-mgmt-databricks>=2.0.0
azure-mgmt-keyvault>=10.2.0
azure-mgmt-storage>=21.0.0

# Infrastructure
# bicep is a CLI tool, not a Python package - install separately if needed

# HTTP and API
requests>=2.31.0
httpx>=0.25.0
fastapi>=0.104.0
uvicorn>=0.24.0

# Database
sqlalchemy>=2.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-asyncio>=0.21.0

# Logging and Monitoring
structlog>=23.2.0
prometheus-client>=0.19.0

# Security
cryptography>=41.0.0
python-jose>=3.3.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3
# uuid is part of Python standard library, no need to install

# Type Checking
mypy>=1.7.0
types-pyyaml>=6.0.12
types-requests>=2.31.0

# Development
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.7.0

